Update Q3Noise.py
Add salt and pepper noise where a pixel is turned into  dark (0) or white (255) with a probability of (p/2) for each  (p to be provided as an input parameter) to 'cap.bmp' image  and then perform denoising by median filtering and mean  filtering. Display the original, noisy, and filtered images.